cleanup types and components
